% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood pressure.R
\name{adjust_DBP}
\alias{adjust_DBP}
\title{Adjust diastolic blood pressure

This function adjusts diastolic blood pressure based on the respondent's diastolic average blood pressure across
six measurements. The adjustment is made using specific correction factors. The adjusted diastolic blood pressure
is returned as a numeric value.}
\usage{
adjust_DBP(BPMDPBPD)
}
\arguments{
\item{BPMDPBPD}{A numeric representing the respondent's diastolic average blood pressure (in mmHg) across six measurements.}
}
\value{
The adjusted diastolic blood pressure as a numeric value.
}
\description{
Adjust diastolic blood pressure

This function adjusts diastolic blood pressure based on the respondent's diastolic average blood pressure across
six measurements. The adjustment is made using specific correction factors. The adjusted diastolic blood pressure
is returned as a numeric value.
}
\details{
The function calculates the adjusted diastolic blood pressure (DBP_adj) based on the value of BPMDPBPD. If
BPMDPBPD is greater than or equal to 0 and less than 996, the adjustment is made using the formula:
DBP_adj = 15.6 + (0.83 * BPMDPBPD). Otherwise, if BPMDPBPD is a non-response value (BPMDPBPD >= 996), the
adjusted diastolic blood pressure is set to NA(b), indicating that the measurement is not available. The adjusted
diastolic blood pressure is returned as the final output.
}
\examples{

# Example: Calculate adjusted diastolic blood pressure for a respondent with an average diastolic blood pressure of 80 mmHg.
adjust_DBP(BPMDPBPD = 80)
# Output: 82

}
