% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood-pressure.R
\name{determine_adjusted_hypertension}
\alias{determine_adjusted_hypertension}
\title{Determine Adjusted Hypertension Status}
\usage{
determine_adjusted_hypertension(SBP_adj, DBP_adj, anymed)
}
\arguments{
\item{SBP_adj}{An integer representing the adjusted systolic blood pressure measurement of the respondent.}

\item{DBP_adj}{An integer representing the adjusted diastolic blood pressure measurement of the respondent.}

\item{anymed}{An integer indicating whether the respondent is on medication for hypertension.
\itemize{
\item 1: Yes
\item 2: No
}}
}
\value{
An integer representing the adjusted hypertension status:
\itemize{
\item 1: High blood pressure (adjusted BP ≥ 140/90 mmHg or on hypertension medication)
\item 2: Normal blood pressure (adjusted BP < 140/90 mmHg and not on hypertension medication)
\item NA(b): Invalid input or non-response
}
}
\description{
This function determines the adjusted hypertension status of a respondent based on their adjusted systolic and diastolic blood pressure measurements and medication usage.
}
\examples{

# Example 1: Determine adjusted hypertension status for a respondent with adjusted systolic BP = 150, adjusted diastolic BP = 95, and on medication.
determine_adjusted_hypertension(SBP_adj = 150, DBP_adj = 95, anymed = 1)
# Output: 1 (High blood pressure due to adjusted systolic BP, adjusted diastolic BP, and medication usage).

# Example 2: Determine adjusted hypertension status for a respondent with adjusted systolic BP = 120, adjusted diastolic BP = 80, and not on medication.
determine_adjusted_hypertension(SBP_adj = 120, DBP_adj = 80, anymed = 2)
# Output: 2 (Normal blood pressure as adjusted BP is below 140/90 mmHg and not on medication).

}
