% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood pressure.R
\name{determine_controlled_adjusted_hypertension}
\alias{determine_controlled_adjusted_hypertension}
\title{Determine Controlled Adjusted Hypertension Status}
\usage{
determine_controlled_adjusted_hypertension(SBP_adj, DBP_adj, anymed)
}
\arguments{
\item{SBP_adj}{An integer representing the adjusted systolic blood pressure measurement of the respondent.}

\item{DBP_adj}{An integer representing the adjusted diastolic blood pressure measurement of the respondent.}

\item{anymed}{An integer indicating whether the respondent is on medication for hypertension.
\itemize{
\item 1: Yes
\item 2: No
}}
}
\value{
An integer representing the controlled adjusted hypertension status:
\itemize{
\item 1: Hypertension not controlled (adjusted BP ≥ 140/90 mmHg or on hypertension medication).
\item 2: Hypertension controlled (adjusted BP < 140/90 mmHg and on hypertension medication).
\item NA(b): Invalid input or non-response.
}
}
\description{
This function determines the controlled adjusted hypertension status of a respondent based on their adjusted systolic and diastolic blood pressure measurements and medication usage.
}
\examples{

# Example 1: Determine controlled adjusted hypertension status for a respondent with adjusted systolic BP = 150, adjusted diastolic BP = 95, and on medication.
determine_controlled_adjusted_hypertension(SBP_adj = 150, DBP_adj = 95, anymed = 1)
# Output: 1 (Hypertension not controlled due to adjusted systolic BP, adjusted diastolic BP, and medication usage).

# Example 2: Determine controlled adjusted hypertension status for a respondent with adjusted systolic BP = 120, adjusted diastolic BP = 80, and on medication.
determine_controlled_adjusted_hypertension(SBP_adj = 120, DBP_adj = 80, anymed = 1)
# Output: 2 (Hypertension controlled as adjusted BP is below 140/90 mmHg and on medication).

}
