% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family history.R
\name{determine_CVD_Family_History}
\alias{determine_CVD_Family_History}
\title{Determine a respondent's cardiovascular disease (CVD) family history based on specific conditions related to heart disease and stroke in their immediate family members.}
\usage{
determine_CVD_Family_History(FMH_11, FMH_12, FMH_13, FMH_14)
}
\arguments{
\item{FMH_11}{An integer indicating whether an immediate family member was diagnosed with heart disease.
1 for "Yes", 2 for "No".}

\item{FMH_12}{A numeric variable representing the youngest age at diagnosis of heart disease in an immediate family member.}

\item{FMH_13}{An integer indicating whether an immediate family member was diagnosed with stroke.
1 for "Yes", 2 for "No".}

\item{FMH_14}{A numeric variable representing the youngest age at diagnosis of stroke in an immediate family member.}
}
\value{
An integer indicating the cardiovascular disease (CVD) family history:
\itemize{
\item 1 for "Yes" if there is a family history of premature CVD (before age 60) based on heart disease and stroke histories.
\item 2 for "No" if there is no such family history.
\item NA(b) if any input contains non-responses or unknown diagnosis ages (997, 998, or 999).
}
}
\description{
Determine a respondent's cardiovascular disease (CVD) family history based on specific conditions related to heart disease and stroke in their immediate family members.
}
\details{
The function evaluates the input variables \code{FMH_11}, \code{FMH_12}, \code{FMH_13}, and \code{FMH_14} to determine the cardiovascular disease (CVD) family history.
\itemize{
\item If \code{FMH_11} is equal to 1 or \code{FMH_13} is equal to 1, it indicates that someone in the family had heart disease or stroke, respectively.
In this case, the function checks the age at diagnosis (\code{FMH_12} for heart disease and \code{FMH_14} for stroke). If the age is greater than or equal to 0 and less than 60,
it represents a diagnosis before age 60, and \code{famheart60} or \code{famstroke60} is set to 1, respectively.
\item If the age is greater than or equal to 60 and less than or equal to 100, it indicates a late diagnosis, and the corresponding \code{famheart60} and \code{famstroke60} are set to 0.
\item If any of the integer inputs contain non-responses (i.e., values greater than 2) or any age input takes non-response values (997, 998, or 999), \code{famCVD60} is set to NA to
indicate that the CVD family history is not available.
\item If any of the conditions for premature CVD (before age 60) are met for either heart disease or stroke, the function sets \code{famCVD60} to 1, indicating a family history of
premature CVD. Otherwise, \code{famCVD60} is set to 2, representing no family history of premature CVD.
}
}
\examples{

# Example: Determine CVD family history for a respondent with a family member diagnosed with heart disease at age 50.
determine_CVD_Family_History(FMH_11 = 1, FMH_12 = 50, FMH_13 = 2, FMH_14 = 6)
# Output: 1 (CVD family history is "Yes" due to a family member's premature heart disease).

}
