% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\name{determine_inclusive_diabetes}
\alias{determine_inclusive_diabetes}
\title{Determine Inclusive Diabetes Status}
\usage{
determine_inclusive_diabetes(diab_m, diab_drug, CCC_51)
}
\arguments{
\item{diab_m}{An integer indicating whether the respondent has diabetes based on HbA1c level. 1 for "Yes", 2 for "No".}

\item{diab_drug}{An integer indicating whether the respondent takes diabetes drugs. 1 for "Yes", 0 for "No".}

\item{CCC_51}{An integer indicating whether the respondent has diabetes in general. 1 for "Yes", 2 for "No".}
}
\value{
An integer indicating the inclusive diabetes status. If any of the input parameters (diab_m, diab_drug, CCC_51)
are non-response values (not 1 or 2), the inclusive diabetes status will be NA(b). If any of the input parameters
is equal to 1, indicating a positive response, the inclusive diabetes status will be 1 ("Yes"). Otherwise, if all
input parameters are 2 ("No"), the inclusive diabetes status will be 2 ("No").
}
\description{
This function evaluates various factors to determine the inclusive diabetes status of a respondent.
}
\examples{

# Example: Determine the inclusive diabetes status for a respondent who has diabetes based on HbA1c.
determine_inclusive_diabetes(diab_m = 1, diab_drug = 2, CCC_51 = 2)
# Output: 1 (Inclusive diabetes status is "Yes" due to having diabetes based on HbA1c).

# Example: Determine the inclusive diabetes status for a respondent who does not have diabetes.
determine_inclusive_diabetes(diab_m = 2, diab_drug = 2, CCC_51 = 2)
# Output: 2 (Inclusive diabetes status is "No" as the respondent does not have diabetes based on any of the factors).

}
