% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{find_week_accelerometer_average}
\alias{find_week_accelerometer_average}
\title{Calculate the average minutes of exercise per day for week-long accelerometer data.

This function calculates the average minutes of exercise per day across a week of accelerometer data. It takes seven
parameters, each representing the minutes of exercise on a specific day (Day 1 to Day 7) of accelerometer measurement.
The function computes the average of these values to obtain the average minutes of exercise per day.}
\usage{
find_week_accelerometer_average(
  AMMDMVA1,
  AMMDMVA2,
  AMMDMVA3,
  AMMDMVA4,
  AMMDMVA5,
  AMMDMVA6,
  AMMDMVA7
)
}
\arguments{
\item{AMMDMVA1}{A numeric representing minutes of exercise on Day 1 of accelerometer measurement.}

\item{AMMDMVA2}{A numeric representing minutes of exercise on Day 2 of accelerometer measurement.}

\item{AMMDMVA3}{A numeric representing minutes of exercise on Day 3 of accelerometer measurement.}

\item{AMMDMVA4}{A numeric representing minutes of exercise on Day 4 of accelerometer measurement.}

\item{AMMDMVA5}{A numeric representing minutes of exercise on Day 5 of accelerometer measurement.}

\item{AMMDMVA6}{A numeric representing minutes of exercise on Day 6 of accelerometer measurement.}

\item{AMMDMVA7}{A numeric representing minutes of exercise on Day 7 of accelerometer measurement.}
}
\value{
A numeric representing the average minutes of exercise per day across a week of accelerometer use.
}
\description{
Calculate the average minutes of exercise per day for week-long accelerometer data.

This function calculates the average minutes of exercise per day across a week of accelerometer data. It takes seven
parameters, each representing the minutes of exercise on a specific day (Day 1 to Day 7) of accelerometer measurement.
The function computes the average of these values to obtain the average minutes of exercise per day.
}
\details{
The function calculates the average minutes of exercise per day by taking the mean of the seven input parameters.
}
\examples{

# Example: Calculate the average minutes of exercise per day for a week of accelerometer data.
find_week_accelerometer_average(30, 40, 25, 35, 20, 45, 50)
# Output: 35 (The average minutes of exercise per day across the week is 35 minutes.)

}
