% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_ace_inhibitor}
\alias{is_ace_inhibitor}
\title{Determine if a CHMS respondent's medication is an ACE inhibitor.

This function checks if a given medication for a CHMS respondent belongs to the ACE inhibitor drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.}
\usage{
is_ace_inhibitor(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A numeric, 1 if medication is in the ACE inhibitor class and 0 if it is not.
}
\description{
Determine if a CHMS respondent's medication is an ACE inhibitor.

This function checks if a given medication for a CHMS respondent belongs to the ACE inhibitor drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.
}
\details{
This function uses the \code{startsWith} function to identify ACE inhibitors based on their ATC codes, which
typically start with "C09". If the ATC code matches the pattern and the medication was taken within the last
month (NPI_25B <= 4), the medication is considered an ACE inhibitor and the function returns TRUE.
Otherwise, it returns FALSE.
}
