% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_any_antiHTN_med}
\alias{is_any_antiHTN_med}
\title{Determine if a CHMS respondent's medication is any anti-hypertensive drug.

This function checks if a given medication for a CHMS respondent belongs to any anti-hypertensive drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.}
\usage{
is_any_antiHTN_med(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A logical, TRUE if medication is in any anti-hypertensive drug class and FALSE if it is not.
}
\description{
Determine if a CHMS respondent's medication is any anti-hypertensive drug.

This function checks if a given medication for a CHMS respondent belongs to any anti-hypertensive drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.
}
\details{
This function first identifies any anti-hypertensive drugs based on their ATC codesm using the regular expression
'C0\link{2, 3, 7, 8, 9}' which matches ATC codes that start with "C02", "C03", "C07", "C08", and "C09". Specific sub-codes
'C07AA07', 'C07AA12', 'C07AG02', 'C03BA08', 'C03CA01', and 'C02KX01' are excluded from the class. If the ATC
code matches the pattern and is not in the exclusion list, and the medication was taken within the last month
(NPI_25B <= 4), the medication is considered an anti-hypertensive drug, and the function returns TRUE.
Otherwise, it returns FALSE.
}
