% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_beta_blocker}
\alias{is_beta_blocker}
\title{Determine if a CHMS respondent's medication is a beta blocker.

This function checks if a given medication for a CHMS respondent belongs to the beta blocker drug class.
The specific conditions for identifying a beta blocker are based on Anatomical Therapeutic Chemical (ATC) codes
and the time when the medication was last taken.}
\usage{
is_beta_blocker(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A logical, TRUE if medication is in the beta blocker class and FALSE if it is not.
}
\description{
Determine if a CHMS respondent's medication is a beta blocker.

This function checks if a given medication for a CHMS respondent belongs to the beta blocker drug class.
The specific conditions for identifying a beta blocker are based on Anatomical Therapeutic Chemical (ATC) codes
and the time when the medication was last taken.
}
\details{
This function uses the \code{startsWith} function to identify beta blockers based on their ATC codes, which
typically start with "C07". Additionally, specific sub-codes 'C07AA07', 'C07AA12', and 'C07AG02' are excluded
from the beta blocker class. If the ATC code matches the pattern and is not in the exclusion list, and the
medication was taken within the last month (NPI_25B <= 4), the medication is considered a beta blocker, and the function
returns TRUE. Otherwise, it returns FALSE.
}
