% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_diabetes_drug}
\alias{is_diabetes_drug}
\title{Determine if a CHMS respondent's medication is a diabetes drug.

This function checks if a given medication for a CHMS respondent belongs to the diabetes drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.}
\usage{
is_diabetes_drug(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A numeric, 1 if medication is in the diabetes drug class and 0 if it is not.
}
\description{
Determine if a CHMS respondent's medication is a diabetes drug.

This function checks if a given medication for a CHMS respondent belongs to the diabetes drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.
}
\details{
This function uses the \code{startsWith} function to identify diabetes drugs based on their ATC codes, which
typically start with "A10". If the ATC code matches the pattern and the medication was taken within the last
month (NPI_25B <= 4), the medication is considered a diabetes drug, and the function returns TRUE.
Otherwise, it returns FALSE.
}
