% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_diuretic}
\alias{is_diuretic}
\title{Determine if a CHMS respondent's medication is a diuretic.

This function checks if a given medication for a CHMS respondent belongs to the diuretic drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.}
\usage{
is_diuretic(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A numeric, 1 if medication is in the diuretic class and 0 if it is not.
}
\description{
Determine if a CHMS respondent's medication is a diuretic.

This function checks if a given medication for a CHMS respondent belongs to the diuretic drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.
}
\details{
This function uses the \code{startsWith} function to identify diuretics based on their ATC codes, which
typically start with "C03". Additionally, specific sub-codes 'C03BA08' and 'C03CA01' are excluded from the
diuretic class. If the ATC code matches the pattern and is not in the exclusion list, and the medication was
taken within the last month (NPI_25B <= 4), the medication is considered a diuretic, and the function
returns TRUE. Otherwise, it returns FALSE.
}
