% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\name{low_drink_score_fun}
\alias{low_drink_score_fun}
\title{Calculate the low drink score for a CHMS respondent based on alcohol consumption.}
\usage{
low_drink_score_fun(CLC_SEX, ALCDWKY)
}
\arguments{
\item{CLC_SEX}{An integer indicating the respondent's sex. 1 for male, 2 for female.}

\item{ALCDWKY}{An integer representing the number of standard drinks consumed by the respondent in a week.}
}
\value{
An integer representing the low drink score:
\itemize{
\item 1 for "Low risk" (0 points).
\item 2 for "Marginal risk" (1-2 points).
\item 3 for "Medium risk" (3-4 points).
\item 4 for "High risk" (5-9 points).
\item NA for invalid input or non-response values.
}
}
\description{
Calculate the low drink score for a CHMS respondent based on alcohol consumption.
}
\details{
The low risk drinking score is based on the scoring system in
Canada's Low-Risk Alcohol Drinking Guideline. The score is divided into two
steps. Step 1 allocates points based on sex and the number of drinks
that you usually have each week. In step 2, one point will be awarded for
each item that is true related to drinking habits. The total score is
obtained from adding the points in step 1 and step 2.
}
\note{
Step 2 is not included in this function because the questions in
step 2 are not asked in any of the CHMS cycles. The score is only based on
step 1.
}
\examples{

# Example: Determine low drink score for a male respondent who consumed 3 standard drinks per week.
low_drink_score_fun(CLC_SEX = 1, ALCDWKY = 3)
# Output: 1 (Low risk due to consuming 3 standard drinks per week).

}
