% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\name{low_drink_score_fun1}
\alias{low_drink_score_fun1}
\title{Calculate the low drink score for a respondent based on alcohol consumption (with former and never categories included).}
\usage{
low_drink_score_fun1(CLC_SEX, ALCDWKY, ALC_17, ALC_11)
}
\arguments{
\item{CLC_SEX}{An integer indicating the respondent's sex. 1 for male, 2 for female.}

\item{ALCDWKY}{An integer representing the number of standard drinks consumed by the respondent in a week.}

\item{ALC_17}{An integer indicating whether the respondent ever drank in their lifetime.
1 for "Yes", 2 for "No".}

\item{ALC_11}{An integer indicating whether the respondent drank alcohol in the past year.
1 for "Yes", 2 for "No".}
}
\value{
An integer representing the low drink score:
\itemize{
\item 1 - Low risk - never drank (0 points)
\item 2 - Low risk - former drinker (0 points)
\item 3 - Marginal risk (1-2 points)
\item 4 - Medium risk (3-4 points)
\item 5 - High risk (5-9 points)
\item NA if any input contains non-response values or is invalid.
}
}
\description{
Calculate the low drink score for a respondent based on alcohol consumption (with former and never categories included).
}
\details{
The low risk drinking score is based on the scoring system in
Canada's Low-Risk Alcohol Drinking Guideline. The score is divided into two
steps. Step 1 allocates points based on sex and the number of drinks
that you usually have each week. In step 2, one point will be awarded for
each item that is true related to drinking habits. The total score is
obtained from adding the points in step 1 and step 2.

This score has two 0 point categories: low risk (never drank) and low risk
(former drinker). The two drinking groups are derived from 'ever had a drink
in lifetime'. 'Ever had a drink in lifetime' is only available in CCHS
2001-2008 and 2015-2018.
}
\note{
Step 2 is not included in this function because the questions in
step 2 are not asked in any of the CHMS cycles. The score is only based on
step 1.
}
\examples{

# Example: Determine low drink score for a male respondent who consumed 3 standard drinks per week and drank within past year.
low_drink_score_fun1(CLC_SEX = 1, ALCDWKY = 3, ALC_17 = 1, ALC_11 = 1)
# Output: 2 (Low risk due to consuming 3 standard drinks per week and having alcohol before).

}
