% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{minperday_to_minperweek}
\alias{minperday_to_minperweek}
\title{Convert minutes per day to minutes per week.

This function takes the average minutes of exercise per day across a week of accelerometer use as an input (\code{MVPA_min}) and
calculates the equivalent minutes of exercise per one week of accelerometer use. The result is returned as a numeric value.}
\usage{
minperday_to_minperweek(MVPA_min)
}
\arguments{
\item{MVPA_min}{A numeric representing the average minutes of exercise per day across a week of accelerometer use.}
}
\value{
A numeric representing the average minutes of exercise per one week of accelerometer use.
}
\description{
Convert minutes per day to minutes per week.

This function takes the average minutes of exercise per day across a week of accelerometer use as an input (\code{MVPA_min}) and
calculates the equivalent minutes of exercise per one week of accelerometer use. The result is returned as a numeric value.
}
\details{
The function simply multiplies the average minutes of exercise per day (\code{MVPA_min}) by 7 to obtain the equivalent
minutes of exercise per one week of accelerometer use.
}
\examples{

# Example: Convert average minutes of exercise per day to minutes per week.
minperday_to_minperweek(35)
# Output: 245 (The equivalent minutes of exercise per one week is 245 minutes.)

}
