% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{pack_years_fun}
\alias{pack_years_fun}
\title{Pack Years Calculation for Smoking History}
\usage{
pack_years_fun(
  SMKDSTY,
  CLC_AGE,
  SMKDSTP,
  SMK_52,
  SMK_31,
  SMK_41,
  SMK_53,
  SMK_23,
  SMK_21,
  SMK_11
)
}
\arguments{
\item{SMKDSTY}{An integer representing the smoking status of the respondent:
\itemize{
\item 1: Daily smoker
\item 2: Occasional smoker (former daily)
\item 3: Occasional smoker (never daily)
\item 4: Former daily smoker (non-smoker now)
\item 5: Former occasional smoker (non-smoker now) who smoked at least 100 cigarettes in their lifetime
\item 6: Non-smoker (never smoked more than 100 cigarettes)
}}

\item{CLC_AGE}{A numeric value representing the respondent's age.}

\item{SMKDSTP}{A numeric value representing the number of years since the respondent stopped smoking completely.}

\item{SMK_52}{A numeric value representing the respondent's age when they first started smoking daily.}

\item{SMK_31}{An numeric representing the number of cigarettes smoked per day for daily smokers.}

\item{SMK_41}{A numeric value representing the number of cigarettes smoked per day for occasional smokers.}

\item{SMK_53}{A numeric value representing the number of years since the respondent quit smoking (for former daily or occasional smokers).}

\item{SMK_23}{A numeric value representing the number of days per month the respondent smoked at least 1 cigarette (for former daily or occasional smokers).}

\item{SMK_21}{A numeric value representing the respondent's age when they first started smoking occasionally.}

\item{SMK_11}{An integer representing whether the respondent has smoked at least 100 cigarettes in their lifetime:
\itemize{
\item 1: Yes
\item 2: No
}}
}
\value{
A numeric value representing the pack years for the respondent's smoking history.
\itemize{
\item If the input parameters are invalid or missing, the function returns NA.
\item For different smoking statuses, the function calculates pack years as follows:
\itemize{
\item Daily smoker: \code{ (CLC_AGE - SMK_52) * (SMK_31 / 20) } (minimum value 0.0137)
\item Occasional smoker (former daily): \code{ ((CLC_AGE - SMK_52 - SMKDSTP) * (SMK_53 / 20)) + (pmax((SMK_41 * (SMK_23 * 12) / 30), 1) * SMKDSTP) }
\item Occasional smoker (never daily): \code{ (pmax((SMK_41 * (SMK_23 * 12) / 30), 1) / 20) * (CLC_AGE - SMK_21) }
\item Former daily smoker (non-smoker now): \code{ pmax(((CLC_AGE - SMK_52 - SMKDSTP) * (SMK_53 / 20)), 0.0137) }
\item Former occasional smoker (non-smoker now) who smoked at least 100 cigarettes lifetime: 0.0137
\item Former occasional smoker (non-smoker now) who have not smoked at least 100 cigarettes lifetime: 0.007
\item Non-smoker: 0
}
}
}
\description{
This function calculates an individual's smoking pack-years based on various CHMS smoking variables. Pack years is a measure used by researchers to quantify lifetime exposure to cigarette use.
}
\examples{

# Example 1: Calculate pack years for a daily smoker who started at age 20 and smokes 30 cigarettes per day.
pack_years_fun(SMKDSTY = 1, CLC_AGE = 40, SMKDSTP = NA, SMK_52 = 20, SMK_31 = 30, SMK_41 = NA, SMK_53 = NA, SMK_23 = NA, SMK_21 = NA, SMK_11 = NA)
# Output: 30 (pack years)

# Example 2: Calculate pack years for a former occasional smoker (non-smoker now) who smoked at least 100 cigarettes in their lifetime.
pack_years_fun(SMKDSTY = 5, CLC_AGE = 50, SMKDSTP = 40, SMK_52 = 18, SMK_31 = NA, SMK_41 = 15, SMK_53 = NA, SMK_23 = 3, SMK_21 = 25, SMK_11 = 1)
# Output: 0.0137 (pack years)

}
\seealso{
https://big-life-lab.github.io/cchsflow/reference/pack_years_fun.html
}
